This might help.. slightly
Turns out its not completely fixed. For some reason it detects a change
when there is a repeater in front of the current a certain distance
from the sign, but if there is a repeater and then exactly 2 wires in
front of the sign, it will always work. This is so hank, but its better
than before